Make the search-input extension point a generic component slot
Replace the hardwired SearchModeSelect mount with the component-registry
slot 'element.listing.search.slots.prefix' (SlotRenderer, same pattern as
asset.tree.node.meta). SearchModeSelect becomes a self-contained default
entry registered by the element module: it reads the listing context via
hooks, owns the mode-change side effects, and renders nothing when no
modes are registered. Any bundle can now put other controls beside the
listing search input without touching the search-mode contract.
